# Conflicts:
#	buildSrc/src/main/java/io/micronaut/guides/feature/SecretsManager.java
#	buildSrc/src/main/resources/pom.xml
#	guides/micronaut-aws-lambda-eventbridge-event/micronaut-aws-lambda-eventbridge-event.adoc
#	guides/micronaut-aws-lambda-s3-event/micronaut-aws-lambda-s3-event.adoc

+++ guides/micronaut-rabbitmq/analytics/java/src/test/java/example/micronaut/AnalyticsServiceTest.java
@@ -27,8 +27,8 @@ public class AnalyticsServiceTest {
         List<BookAnalytics> analytics = analyticsService.listAnalytics();
         assertEquals(2, analytics.size());
 
-        assertEquals(3, findBookAnalytics(b1, analytics).getCount().intValue());
-        assertEquals(1, findBookAnalytics(b2, analytics).getCount().intValue());
+        assertEquals(3, findBookAnalytics(b1, analytics).getCount());
+        assertEquals(1, findBookAnalytics(b2, analytics).getCount());
     }
 
     private BookAnalytics findBookAnalytics(Book b, List<BookAnalytics> analytics) {
